# api_documentation/Core.swagger_collection.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Deep Lynx",
...
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/containers/{container-id}": {
      "put": {
        "description": "Updates the container. This will fail if a container already exists with the proposed updated name.",
        "summary": "Update",
        "tags": [
          "Containers"
        ],
        "operationId": "Update",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Fetch container by ID.",
        "summary": "Fetch by ID",
        "tags": [
          "Containers"
        ],
        "operationId": "FetchbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Container. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive",
        "tags": [
          "Containers"
        ],
        "operationId": "Archive",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/containers/{container-id}/metatype_relationships/{relationship-id}/keys": {
      "post": {
        "description": "Creates a new key for a type relationship. Keys consist of a unique key name (unique to the type not as a whole), key type, default values, and allowed values. Of those, only the first two are required.\n\nThe `dataType` field accepts only one of the following values: number, string, date, boolean, enumeration. \n\nThe fields `defaultValue` and `options` will only accept an array of the following types: string, boolean, number, float.",
        "summary": "New Type Relationship Key",
        "tags": [
          "Metatype Relationship Keys"
        ],
        "operationId": "NewTypeRelationshipKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewTypeRelationshipKeyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatype_relationships/{relationship-id}/keys/{key-id}": {
      "get": {
        "description": "Retrieve a single key for a Type Relationship by id.",
        "summary": "Get Type Relationship Key",
        "tags": [
          "Metatype Relationship Keys"
        ],
        "operationId": "GetTypeRelationshipKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "Updates a Type Relationship key. The update must follow the same format as creation.",
        "summary": "Update Type Relationship Key",
        "tags": [
          "Metatype Relationship Keys"
        ],
        "operationId": "UpdateTypeRelationshipKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateTypeRelationshipKeyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Type Relationship Key. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive Type Relationship Key",
        "tags": [
          "Metatype Relationship Keys"
        ],
        "operationId": "ArchiveTypeRelationshipKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{id}/metatypes_relationships/{relationship-id}/keys": {
      "get": {
        "description": "Retrieves all keys for a Type Relationship.",
        "summary": "List Type Relationship Keys",
        "tags": [
          "Metatype Relationship Keys"
        ],
        "operationId": "ListTypeRelationshipKeys",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "relationship-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatype_relationship_pairs": {
      "post": {
        "description": "Create a new Type Relationship Pair. Describes the connection between two Types by connecting them using a Type Relationship and acts as a vehicle for relationship specific keys.\n\nPass in an array for bulk creation.",
        "summary": "New Type Relationship Pair",
        "tags": [
          "Metatype Relationship Pairs"
        ],
        "operationId": "NewTypeRelationshipPair",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewTypeRelationshipPairRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "List all Type Relationship Pairs for current container.",
        "summary": "List Type Relationship Pairs",
        "tags": [
          "Metatype Relationship Pairs"
        ],
        "operationId": "ListTypeRelationshipPairs",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatype_relationship_pairs/{pair-id}": {
      "get": {
        "description": "Retrieves a single Type Relationship Pair.",
        "summary": "Get Type Relationship Pair",
        "tags": [
          "Metatype Relationship Pairs"
        ],
        "operationId": "GetTypeRelationshipPair",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "pair-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Type Relationship Pair. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive Type Pair",
        "tags": [
          "Metatype Relationship Pairs"
        ],
        "operationId": "ArchiveTypePair",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "pair-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateRequest": {
      "title": "UpdateRequest",
      "example": {
        "description": "Final Test Description"
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        }
      },
      "required": [
        "description"
      ]
    },
...
    "Validation": {
      "title": "Validation",
      "example": {
        "regex": "",
        "min": 0,
        "max": 0
      },
      "type": "object",
      "properties": {
        "regex": {
          "type": "string"
        },
        "min": {
          "type": "integer",
          "format": "int32"
        },
        "max": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "regex",
        "min",
        "max"
      ]
    },
...
    "NewTypeRelationshipKeyRequest": {
      "title": "NewTypeRelationshipKeyRequest",
      "example": {
        "name": "File Name",
        "required": true,
        "property_name": "fileName",
        "description": "File Upload Name",
        "data_type": "string",
        "cardinality": 1,
        "validation": {
          "regex": "",
          "min": 0,
          "max": 0
        },
        "unique": true,
        "options": [
          ""
        ],
        "defaultValue": "can be any type"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "required": {
          "type": "boolean"
        },
        "property_name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "data_type": {
          "type": "string"
        },
        "cardinality": {
          "type": "integer",
          "format": "int32"
        },
        "validation": {
          "$ref": "#/definitions/Validation"
        },
        "unique": {
          "type": "boolean"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "defaultValue": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "required",
        "property_name",
        "description",
        "data_type",
        "cardinality",
        "validation",
        "unique",
        "options",
        "defaultValue"
      ]
    },
    "UpdateTypeRelationshipKeyRequest": {
      "title": "UpdateTypeRelationshipKeyRequest",
      "example": {
        "name": "Updated Name",
        "description": "Updated Name Description",
        "data_type": "string",
        "default_value": "optional - data structure here is dependent on dataType value"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "data_type": {
          "type": "string"
        },
        "default_value": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description",
        "data_type",
        "default_value"
      ]
    },
    "NewTypeRelationshipPairRequest": {
      "title": "NewTypeRelationshipPairRequest",
      "example": {
        "name": "Relationship Pair Test",
        "description": "Relationship Pair Test description",
        "origin_metatype_id": "566e3862-6853-45af-8a45-3050576dfb9a",
        "destination_metatype_id": "8bcc0b06-bb44-4176-a581-aa68eafca7f3",
        "metatype_relationship_id": "885e3fca-5066-4be8-a720-63b3c4744e97",
        "relationship_type": "many:many"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "origin_metatype_id": {
          "type": "string"
        },
        "destination_metatype_id": {
          "type": "string"
        },
        "metatype_relationship_id": {
          "type": "string"
        },
        "relationship_type": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description",
        "origin_metatype_id",
        "destination_metatype_id",
        "metatype_relationship_id",
        "relationship_type"
      ]
    },
...
  },
...
}